feat: add support for MCLI_PROFILE
